fix: set mui stuff as state
