+++ src/index.js
@@ -3,6 +3,7 @@ import properties from './object-properties';
 import initSelections from './selections-factory';
 import data from './data';
 import ext from './ext';
+import muiSetup from './mui-setup';
 import { render, teardown } from './table/root';
 import manageData from './table/handle-data';
 
@@ -22,6 +23,7 @@ export default function supernova(env) {
 
       const [pageInfo, setPageInfo] = useState({ top: 0, height: 100 });
       const [tableData, setTableData] = useState();
+      const [muiParameters] = useState(muiSetup(constraints.active));
 
       const selections = initSelections(el);
 
@@ -40,7 +42,7 @@ export default function supernova(env) {
 
       useEffect(() => {
         if (layout && tableData) {
-          render(el, { tableData, setPageInfo, pageInfo, constraints, selections });
+          render(el, { tableData, setPageInfo, pageInfo, constraints, selections, muiParameters });
         }
       }, [tableData, selections.selected]);
     },
